+++ packages/genesys/packages/react-components-core/src/config/createConfig.ts
@@ -20,7 +20,6 @@ export default function createConfigCore(
         locale: locale || defaultLocale,
         validators: { defaultValidators, ...validators },
         components: deepmerge(defaultComponents, components || {}),
-        ...restDefaults,
-        ...extraConfig,
+        ...deepmerge(restDefaults, extraConfig),
     };
 }