+++ packages/genesys/packages/react-native-components/src/input/Button/hooks/useButtonStyles.ts
@@ -1,9 +1,11 @@
 import { TextStyle, ViewStyle } from "react-native";
-import useDefaultStyles from "./useDefaultStyles";
 import { extractTextStyles } from "@peersyst/react-native-utils";
-import { ButtonStyle, ButtonStyles } from "../Button.types";
+import { ButtonRootStyle, ButtonStyle, ButtonTextStyle } from "../Button.types";
 import { useMemo } from "react";
-import { ButtonSize, ButtonVariant } from "@peersyst/react-components-core";
+import useButtonDefaultStyle from "./useButtonDefaultStyle.ts";
+import { useGlobalStyles } from "src/config";
+import { StyleFlags } from "../../../utils";
+import { useMergeStyles } from "../../../hooks";
 
 export interface UseButtonStylesResult {
     textStyle: TextStyle;
@@ -11,77 +13,35 @@ export interface UseButtonStylesResult {
 }
 
 export default function useButtonStyles(
-    style: ButtonStyles,
-    variant: ButtonVariant,
-    size: ButtonSize,
-    disabled: boolean,
-    pressed: boolean,
+    style: ButtonStyle,
     color: string | undefined,
-): UseButtonStylesResult {
-    const { defaultStyles, defaultDisabledStyles, defaultPressedStyles, defaultSizeStyles } =
-        useDefaultStyles(color);
-    const {
-        disabled: { variant: variantDisabledStyles, ...disabledStyles } = {},
-        pressed: { variant: variantPressedStyles, ...pressedStyles } = {},
-        ...styles
-    } = style;
+    {
+        filled,
+        outlined,
+        text,
+        sm,
+        md,
+        lg,
+        pressed,
+        disabled,
+    }: StyleFlags<
+        ButtonStyle,
+        "filled" | "outlined" | "text" | "sm" | "md" | "lg" | "pressed" | "disabled"
+    >,
+): [ButtonRootStyle, ButtonTextStyle] {
+    const defaultStyle = useButtonDefaultStyle(color);
+    const globalStyle = useGlobalStyles("Button");
 
-    const [textStyles, rootStyles] = useMemo(
-        () => extractTextStyles({ ...defaultStyles, ...styles }),
-        [defaultStyles, styles],
-    );
-    const [disabledTextStyles, disabledRootStyles] = useMemo(
-        () => extractTextStyles({ ...defaultDisabledStyles, ...disabledStyles }),
-        [defaultDisabledStyles, disabledStyles],
-    );
+    const mergedStyles = useMergeStyles([defaultStyle, globalStyle, style], {
+        filled,
+        outlined,
+        text,
+        sm,
+        md,
+        lg,
+        pressed,
+        disabled,
+    });
 
-    const [variantDisabledTextStyles, variantDisabledRootStyles] = useMemo(() => {
-        return extractTextStyles({
-            ...defaultDisabledStyles.variant?.[variant],
-            ...variantDisabledStyles?.[variant],
-        });
-    }, [variantDisabledStyles, defaultDisabledStyles, variant]);
-
-    const [pressedTextStyles, pressedRootStyles] = useMemo(
-        () => extractTextStyles({ ...defaultPressedStyles, ...pressedStyles }),
-        [defaultPressedStyles, pressedStyles],
-    );
-
-    const [variantPressedTextStyles, variantPressedRootStyles] = useMemo(() => {
-        return extractTextStyles({
-            ...defaultPressedStyles.variant?.[variant],
-            ...variantPressedStyles?.[variant],
-        });
-    }, [variantPressedStyles, defaultPressedStyles, variant]);
-
-    const [variantTextStyles, variantRootStyles] = useMemo(
-        () =>
-            extractTextStyles({
-                ...defaultStyles.variant?.[variant],
-                ...styles.variant?.[variant],
-            }),
-        [defaultStyles, styles, variant],
-    );
-
-    const [sizeTextStyles, sizeRootStyles] = useMemo(
-        () => extractTextStyles({ ...defaultSizeStyles[size], ...styles[size] }),
-        [defaultSizeStyles, styles, size],
-    );
-
-    const textStyle = {
-        ...textStyles,
-        ...variantTextStyles,
-        ...sizeTextStyles,
-        ...(disabled && { ...disabledTextStyles, ...variantDisabledTextStyles }),
-        ...(pressed && { ...pressedTextStyles, ...variantPressedTextStyles }),
-    };
-    const rootStyle = {
-        ...rootStyles,
-        ...variantRootStyles,
-        ...sizeRootStyles,
-        ...(disabled && { ...disabledRootStyles, ...variantDisabledRootStyles }),
-        ...(pressed && { ...pressedRootStyles, ...variantPressedRootStyles }),
-    };
-
-    return { textStyle, rootStyle };
+    return useMemo(() => extractTextStyles(mergedStyles), [mergedStyles]);
 }