+++ packages/genesys/packages/react-native-components/src/utils/index.ts
@@ -1,1 +1,4 @@
 export * from "./booleanStylesMergeStrategy";
+export * from "./extractFlagStyles";
+export { default as extractFlagStyles } from "./extractFlagStyles";
+export { default as flattenStyles } from "./flattenStyles";