+++ packages/genesys/packages/react-native-utils/src/extractTextStyles.ts
@@ -1,79 +1,43 @@
 import { TextStyle } from "react-native";
 
-type ExtractedTextStyles =
-    | "color"
-    | "fontFamily"
-    | "fontSize"
-    | "fontStyle"
-    | "fontWeight"
-    | "letterSpacing"
-    | "lineHeight"
-    | "textAlign"
-    | "textDecorationLine"
-    | "textDecorationStyle"
-    | "textDecorationColor"
-    | "textShadowColor"
-    | "textShadowOffset"
-    | "textShadowRadius"
-    | "textTransform"
-    | "fontVariant"
-    | "writingDirection"
-    | "textAlignVertical"
-    | "includeFontPadding";
+export const textStyleKeys = [
+    "color",
+    "fontFamily",
+    "fontSize",
+    "fontStyle",
+    "fontWeight",
+    "letterSpacing",
+    "lineHeight",
+    "textAlign",
+    "textDecorationLine",
+    "textDecorationStyle",
+    "textDecorationColor",
+    "textShadowColor",
+    "textShadowOffset",
+    "textShadowRadius",
+    "textTransform",
+    "fontVariant",
+    "writingDirection",
+    "textAlignVertical",
+    "includeFontPadding",
+] as const;
 
-// eslint-disable-next-line @typescript-eslint/ban-types
+export type TextStylesKeys = typeof textStyleKeys[number];
+
 export default function extractTextStyles<T extends TextStyle>(
     styles?: T,
-): [TextStyle, Omit<T, ExtractedTextStyles> | {}] {
-    if (!styles) return [{}, {}];
-    const {
-        color,
-        fontFamily,
-        fontSize,
-        fontStyle,
-        fontWeight,
-        letterSpacing,
-        lineHeight,
-        textAlign,
-        textDecorationLine,
-        textDecorationStyle,
-        textDecorationColor,
-        textShadowColor,
-        textShadowOffset,
-        textShadowRadius,
-        textTransform,
-        fontVariant,
-        writingDirection,
-        textAlignVertical,
-        includeFontPadding,
-        ...rest
-    } = styles;
-    const textStyles = {
-        color,
-        fontFamily,
-        fontSize,
-        fontStyle,
-        fontWeight,
-        letterSpacing,
-        lineHeight,
-        textAlign,
-        textDecorationLine,
-        textDecorationStyle,
-        textDecorationColor,
-        textShadowColor,
-        textShadowOffset,
-        textShadowRadius,
-        textTransform,
-        fontVariant,
-        writingDirection,
-        textAlignVertical,
-        includeFontPadding,
-    };
-    Object.keys(textStyles).forEach(
-        (key) =>
-            textStyles[key as keyof typeof textStyles] === undefined &&
-            delete textStyles[key as keyof typeof textStyles],
-    );
+): [TextStyle, Omit<T, TextStylesKeys>] {
+    if (!styles) return [{}, {} as Omit<T, TextStylesKeys>];
 
-    return [textStyles, rest];
+    return Object.entries(styles).reduce(
+        ([textStyles, rest], [key, value]) => {
+            if (textStyleKeys.includes(key as TextStylesKeys)) {
+                textStyles[key as TextStylesKeys] = value;
+            } else {
+                rest[key as Exclude<keyof T, TextStylesKeys>] = value;
+            }
+            return [textStyles, rest];
+        },
+        [{} as TextStyle, {} as Omit<T, TextStylesKeys>],
+    );
 }