+++ packages/react-native-styled/src/types.ts
@@ -2,9 +2,14 @@ import { Theme } from "@peersyst/react-native-styled";
 import { ScaledSize } from "react-native";
 import { EdgeInsets } from "react-native-safe-area-context";
 
-export type SX<TStyle> = (theme: Theme) => TStyle;
+export type SX<TStyle> = (p: {
+    theme: Theme;
+    dimensions: ScaledSize;
+    safeAreaInsets: EdgeInsets;
+}) => TStyle;
 
-// eslint-disable-next-line @typescript-eslint/ban-types
 export type StyledFunction<P extends { style?: P["style"] }, E = {}> = (
-    p: { theme: Theme } & { dimensions: ScaledSize; safeAreaInsets: EdgeInsets } & E,
+    p: { theme: Theme; dimensions: ScaledSize; safeAreaInsets: EdgeInsets } & E,
 ) => P["style"];
+
+export type StyledComponentProps<S> = { sx?: SX<S>; style?: S };