+++ packages/genesys/packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Versions
 
+## 1.2.6
+
+Dependencies update
 ## 1.2.5
 
 Dependencies update