+++ packages/genesys/packages/react-native-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Versions
 
+## 1.1.0
+
+Dependencies update
+
 ## 0.0.4
 
 Dependencies update