+++ packages/react-native-styled/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 2.1.0
+
+[styled] Improve useMemos
+[styled] Loosen props given in the styled fn
+
 ## 1.2.0
 
 - Styled also gives safeAreaInsets now.